# uni-borrow-backend/openapi.yaml
openapi: 3.0.3
info:
  title: School Equipment Loan Management API
  description: REST API for managing school equipment loans with role-based access control
  version: 1.0.0
  contact:
    name: API Support
    email: support@schoolequipment.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.schoolequipment.com/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BorrowRequest:
      type: object
      properties:
        id:
          type: string
          example: "req1"
        equipmentId:
          type: string
          example: "eq1"
        userId:
          type: string
          example: "1"
        userName:
          type: string
          example: "John Doe"
        status:
          type: string
          enum: [pending, approved, rejected, returned]
          example: "pending"
        requestDate:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
        approvedDate:
          type: string
          format: date-time
          nullable: true
          example: "2025-01-15T14:20:00Z"
        returnDate:
          type: string
          format: date-time
          nullable: true
          example: "2025-01-20T09:15:00Z"
        approvedBy:
          type: string
          nullable: true
          example: "Jane Smith"
        quantity:
          type: integer
          minimum: 1
          example: 1
        notes:
          type: string
          nullable: true
          example: "Needed for project work"
      required:
        - id
        - equipmentId
        - userId
        - userName
        - status
        - requestDate
        - quantity
...
    CreateBorrowRequest:
      type: object
      properties:
        equipmentId:
          type: string
          example: "eq1"
        quantity:
          type: integer
          minimum: 1
          example: 1
        notes:
          type: string
          nullable: true
          example: "Need for upcoming presentation"
      required:
        - equipmentId
        - quantity
    UpdateRequestStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected, returned]
          example: "approved"
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid credentials"
        message:
          type: string
          example: "The email or password provided is incorrect"
        statusCode:
          type: integer
          example: 401
paths:
...
  /requests:
    get:
      tags:
        - Borrow Requests
      summary: List borrow requests
      description: Retrieve borrow requests based on user role (students see their own, staff/admin see all)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, returned]
          description: Filter by status
          example: "pending"
        - name: userId
          in: query
          schema:
            type: string
          description: Filter by user ID (staff/admin only)
          example: "1"
        - name: equipmentId
          in: query
          schema:
            type: string
          description: Filter by equipment ID
          example: "eq1"
      responses:
        '200':
          description: Requests retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BorrowRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Borrow Requests
      summary: Create borrow request
      description: Submit a new equipment borrow request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBorrowRequest'
      responses:
        '201':
          description: Request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRequest'
        '400':
          description: Invalid input or insufficient equipment availability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /requests/{id}:
    get:
      tags:
        - Borrow Requests
      summary: Get request by ID
      description: Retrieve specific borrow request details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "req1"
      responses:
        '200':
          description: Request details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRequest'
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /requests/{id}/status:
    patch:
      tags:
        - Borrow Requests
      summary: Update request status
      description: Approve, reject, or mark as returned (Staff/Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "req1"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequestStatusRequest'
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRequest'
        '400':
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Staff/Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...